Material 3 (#988)
* use real material dialogs

* Fixes for tristate preference

* use material 1.5.0-alpha03

* Material 3 start

* More Material 3 changes

Updates to color primary for many themes
Using material dialog in more places
Instead of toolbar elevation, changing the color of the toolbar instead

* fix toolbar bg coloring

* More changes and fixes

fixed dialog width when using forced side nav

* update more color primaries

* Fix toolbar title alpha

* more updates

Updates to many fonts (still more to come)
Update to tracker dialogs to true material

* More dialogs updated

Using extensions for some old functionality in material dialog library

* Removed final use of mat dialogs

* Fully removed afollestad.material-dialogs

you will be missed tbh

* fix overflow width back to normal + bg color change for it

* Fix RTL for reader slider

* Fix RTL in general

This way I can revert back the slider changes.

thanks @arkon

* Android 12/SDK 31 Target

* restore ManageCategoryDialog's postive button action

no longer dismisses if there's errors with something

* Use disable items helper in MultiListMatPreference

* Change statelist and ripple for update button in ext page

* Update runtime-ktx and appcompat to alpha for A12 Target

* Removing hyperion as its no longer compatible with A12 target

Sorry Carlos, I never really never used it, even though its neat

Maybe if the library gets an update, but it hasn't had once in Jan

* Fix not being able to install extensions under A12 target

* Fixed status bar color from going back to extensions list from ext settings

* modernizing the action mode toolbar

* tinting overlay background with a hint of secondary

* Remove use of BuildCompat

since `isAtLeastS` will be deprecated soon enough

* Change startInstalling method in updateservice to only run on A12

* Revert dialog behavior for which category to update on swipe refresh

* Adding primaryInverse

used for snackbar action + reader progress when using a different background color from theme (i.e. yin theme with a white reader background)

* More subtle coloring of the manga details backdrop

theres still a tint but its next to unnoticeable now

* Remove colorDownloadBadge for colorTertiary

* Using more rounded indicator for tab chips

* removed randomly extra line in styles

* remove suggested_date_ string since no longer used

* Update chucker version to 3.5.2

to support a12 target

* No longer coloring the highlighted bottom nav text

* many fixes to preference dialogs

along with making the tri state pref match other choice pref dialogs

* Switch tabs style to use M3

No real change as far as i can see

* setting expedited for job

as needed for bigger job in A12

* Changing active tab colors again

changing to match other google apps (gmail, phone, photos) instead of trying to match other google apps (clock)

Also upping the brightness of colorSecondaryVariant for many themes

* Reverting back to target a11

Not worth trying to fix a12 issues with background jobs

* More appropriate colors for colorSecondaryVariant in light themes

* more updates to the active tab colors

* Using sliders + switch material for the reader color filters

Also updating the tints of switches

Will fix #1004

* Changing library badge style + increasing size of grid items

Library badge is now inside the grid, so less padding is needed for the grid item, thus its bigger now
Since its bigger now, the search toolbar is now wider to match its width

* Fix incognito badge not showing until title is showing on manga details

* revert work runtime to 2.6.0

* fix library unread badge for list mode

* fix text colors for download tooltips

* Material3 popup in webview activity

* removing more uses of MaterialComponents for M3

* update header card always show category name

* fix missing text appearance

* Cleanup SearchController

* clean up SwitchPreferenceCategory

* optimize imports of SearchController

* Use SwitchPreferenceCompat in SettingsExtensionsController

* Use of material 3 buttons in manga details library/tracking

* Option to theme manga details page buttons based on the cover

Backdrop tint color is now based on the covers dominant cover instead of the vibrant color, that is now used for the buttons

Also prefetching those colors when loading the covers from other screens

* removing stroke on the cover in details

* improvements to the navigation icon + blue theme improvements

+ styling the swipe refresh to the magna theme

* general cleanup/fixes when it comes to light status/nav bar

also general fixes to webview theme switching

* styles cleanup

* improvements to the library grid selector

* random refactoring thats honestly unrelated to this branch

im nearing the end of development for m3 now

* Details' bookmarks now themed by cover

* Use Material3 for primary buttons

* Use new mat dialog for filter scanlators sheet

* Updates to themed details refresh icon

* fix toolbar title being centered in searchactivity

* Updates to library badges

Now slightly outside of the covers again

* Global search badge updated to match library badge

* Updates to the manga details theme colors to be closer to the original suggested color

Also the backdrop tint will no longer be used as much if it makes for a bad contrast between it and the title/subtitle text

* Updates to the monet themes colorPrimaryVariant

* more updates to cover theme colors

tries to pick between dominant, vibrant, and muted to find something colorful to use

* Updated to material alpha04

With it brings:
M3 Nav Rail
Updated dialogs
Updated sheets theme

* Delete side_nav_item_selector.xml

* Update colorPrimVariant for light themes

* remove elevation in theme preview item

* various updates to imports

* Update ControllerExtensions.kt

* Fix extensions/ dl queue title not being centered

* more improvements to bestColor for covers

* Use attr for styles of textviews

+ style cleanup

* Use cross in filter scanlator groups to show its filtered

* Support delta movement of moving manga categories

Tristate to manage add/remove without removing categories categories

also some refactoring of TriStateCheckBox
closes #644

* remove stroke on category hopper

* update to details theming in light theme

* Fix some text appearances

* Fixes to text that shows when some categories are shared between selected manga

* Use Material3 style for the final remaining elements

this is slider tooltip, chips, text buttons, editText, tracker cardview

* making category hopper slightly bigger

* fixed dialog divider color for edit manga and tri state

* fixed top padding in manga category dialog

* Update dim for full cover for devices under A12

* More improvements to the logic for what text shows on the move categories button

* Fixed slider tooltip color and text size

as well as made the inactive ticks less visible

* fixed light nav bar in bottom sheets for older devices

* Fix reader nav bar colors for android 10 and below

for android 8 and below the reader nav bar is now black in light themes
also removed zhanghai.android.systemuihelper lib

* set tri checkbox background ripple to off for dialogs

* fixed misuse of textAppearanceBodySmall

* make lightStatusBar a local val in readeractivity

* when manga is in the default category, text should always say "move to" instead "add to" when change categories

* Fixed copy and logic for when trying to remove tracker while offline

(cherry picked from commit 9b0e0fcfe65c174829e8d6f73cb338eec8eb156e)

+++ app/src/main/java/eu/kanade/tachiyomi/ui/category/CategoryController.kt
@@ -5,7 +5,6 @@ import android.view.LayoutInflater
 import android.view.View
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.afollestad.materialdialogs.MaterialDialog
 import com.google.android.material.snackbar.BaseTransientBottomBar
 import com.google.android.material.snackbar.Snackbar
 import eu.davidea.flexibleadapter.FlexibleAdapter
@@ -14,6 +13,7 @@ import eu.kanade.tachiyomi.databinding.CategoriesControllerBinding
 import eu.kanade.tachiyomi.ui.base.controller.BaseController
 import eu.kanade.tachiyomi.ui.category.CategoryPresenter.Companion.CREATE_CATEGORY_ORDER
 import eu.kanade.tachiyomi.ui.main.MainActivity
+import eu.kanade.tachiyomi.util.system.materialAlertDialog
 import eu.kanade.tachiyomi.util.system.toast
 import eu.kanade.tachiyomi.util.view.liftAppbarWith
 import eu.kanade.tachiyomi.util.view.snack
@@ -124,13 +124,13 @@ class CategoryController(bundle: Bundle? = null) :
     }
 
     override fun onItemDelete(position: Int) {
-        MaterialDialog(activity!!)
-            .title(R.string.confirm_category_deletion)
-            .message(R.string.confirm_category_deletion_message)
-            .positiveButton(R.string.delete) {
+        activity!!.materialAlertDialog()
+            .setTitle(R.string.confirm_category_deletion)
+            .setMessage(R.string.confirm_category_deletion_message)
+            .setPositiveButton(R.string.delete) { _, _ ->
                 deleteCategory(position)
             }
-            .negativeButton(android.R.string.no)
+            .setNegativeButton(android.R.string.cancel, null)
             .show()
     }
